+++ src/app/controllers/AppointmentController.js
@@ -1,13 +1,9 @@
-import { isBefore, subHours } from 'date-fns';
-
 import User from '../models/User';
 import Appointment from '../models/Appointment';
 import File from '../models/File';
 
-import CancellationMail from '../jobs/CancellationMail';
-import Queue from '../../lib/Queue';
-
 import CreateAppointmentService from '../services/CreateAppointmentService';
+import CancelAppointmentService from '../services/CancelAppointmentService';
 
 /**
  * Apontamentos do usuario/cliente
@@ -54,40 +50,10 @@ class AppointmentController {
   }
 
   async delete(req, resp) {
-    const appointment = await Appointment.findByPk(req.params.id, {
-      include: [
-        {
-          model: User,
-          as: 'provider',
-          attributes: ['name', 'email'],
-        },
-        {
-          model: User,
-          as: 'user',
-          attributes: ['name'],
-        },
-      ],
+    const appointment = await CancelAppointmentService.run({
+      provider_id: req.params.id,
+      user_id: req.userId,
     });
-
-    if (appointment.user_id !== req.userId) {
-      return resp.status(401).json({
-        error: "You don't have permission to cancel this appointment",
-      });
-    }
-
-    const dateWithSub = subHours(appointment.date, 2);
-
-    if (isBefore(dateWithSub, new Date())) {
-      return resp.status(401).json({
-        error: 'You can only cancel appointments 2 hours in advance',
-      });
-    }
-
-    appointment.canceled_at = new Date();
-
-    await appointment.save();
-
-    Queue.add(CancellationMail.key, { appointment });
 
     return resp.json(appointment);
   }
